Move attempts to attempts folder
